LUCENE-9705: Final clean-up and entry in CHANGES.txt
